consumer decomposing & log level changes
