+++ vent/types/sql_table.go
@@ -65,7 +65,6 @@ const (
 	SQLColumnLabelEventType   = "_eventtype"
 	SQLColumnLabelBlockHeader = "_blockheader"
 	SQLColumnLabelTxType      = "_txtype"
-	SQLColumnLabelTxExec      = "_txexecutions"
 	SQLColumnLabelEnvelope    = "_envelope"
 	SQLColumnLabelEvents      = "_events"
 	SQLColumnLabelResult      = "_result"
